* Speaker inference.
Inference is performed on an arbitrary dataset that must consist of
image pairs. Captions that describe those differences are generated
using a previously-trained speaker model.

I created new files (run_inference.py, src/inference.py and
src/data_infer.py) for this task, to prevent the introduction of bugs in
the existing code. The only modification was to add a special purpose
`inference` method to `speaker.py`.